+++ src/database/src/DetectsLostConnections.php
@@ -43,7 +43,6 @@ trait DetectsLostConnections
             'Name or service not known',
             'ORA-03114',
             'Packets out of order. Expected',
-            'There is no active transaction',
         ]);
     }
 }