+++ src/grpc/src/Parser.php
@@ -13,6 +13,8 @@ namespace Hyperf\Grpc;
 
 use Google\Protobuf\GPBEmpty;
 use Google\Protobuf\Internal\Message;
+use Google\Rpc\Status;
+use Swoole\Http\Response;
 use swoole_http2_response;
 
 class Parser
@@ -35,18 +37,7 @@ class Parser
 
     public static function serializeMessage($data)
     {
-        if ($data === null) {
-            $data = new GPBEmpty();
-        }
-        if (method_exists($data, 'encode')) {
-            $data = $data->encode();
-        } elseif (method_exists($data, 'serializeToString')) {
-            $data = $data->serializeToString();
-        } elseif (method_exists($data, 'serialize')) {
-            /** @noinspection PhpUndefinedMethodInspection */
-            $data = $data->serialize();
-        }
-        return self::pack((string) $data);
+        return self::pack(self::serializeUnpackedMessage($data));
     }
 
     public static function deserializeMessage($deserialize, string $value)
@@ -54,21 +45,8 @@ class Parser
         if (empty($value)) {
             return null;
         }
-        $value = self::unpack($value);
 
-        if (is_array($deserialize)) {
-            [$className, $deserializeFunc] = $deserialize;
-            /** @var \Google\Protobuf\Internal\Message $object */
-            $object = new $className();
-            if ($deserializeFunc && method_exists($object, $deserializeFunc)) {
-                $object->{$deserializeFunc}($value);
-            } else {
-                // @noinspection PhpUndefinedMethodInspection
-                $object->mergeFromString($value);
-            }
-            return $object;
-        }
-        return call_user_func($deserialize, $value);
+        return self::deserializeUnpackedMessage($deserialize, self::unpack($value));
     }
 
     /**
@@ -94,6 +72,59 @@ class Parser
         $reply = self::deserializeMessage($deserialize, $data);
         $status = (int) ($response->headers['grpc-status'] ?? 0);
         return [$reply, $status, $response];
+    }
+
+    /**
+     * @param Response
+     * @param mixed $response
+     */
+    public static function statusFromResponse($response): ?Status
+    {
+        $detailsEncoded = $response->headers['grpc-status-details-bin'] ?? '';
+
+        if (! $detailsEncoded || ! $detailsBin = base64_decode($detailsEncoded, true)) {
+            return null;
+        }
+        return self::deserializeUnpackedMessage([Status::class, ''], $detailsBin);
+    }
+
+    public static function statusToDetailsBin(Status $status): string
+    {
+        return base64_encode(self::serializeUnpackedMessage($status));
+    }
+
+    private static function deserializeUnpackedMessage($deserialize, string $unpacked)
+    {
+        if (is_array($deserialize)) {
+            [$className, $deserializeFunc] = $deserialize;
+            /** @var \Google\Protobuf\Internal\Message $object */
+            $object = new $className();
+            if ($deserializeFunc && method_exists($object, $deserializeFunc)) {
+                $object->{$deserializeFunc}($unpacked);
+            } else {
+                // @noinspection PhpUndefinedMethodInspection
+                $object->mergeFromString($unpacked);
+            }
+            return $object;
+        }
+        return call_user_func($deserialize, $unpacked);
+    }
+
+    private static function serializeUnpackedMessage($data): string
+    {
+        if ($data === null) {
+            $data = new GPBEmpty();
+        }
+        if (method_exists($data, 'encode')) {
+            $data = $data->encode();
+        } elseif (method_exists($data, 'serializeToString')) {
+            $data = $data->serializeToString();
+        } elseif (method_exists($data, 'serialize')) {
+            /** @noinspection PhpUndefinedMethodInspection */
+            $data = $data->serialize();
+        }
+
+        return (string) $data;
     }
 
     private static function isinvalidStatus(int $code)