+++ src/kafka/src/ConsumerManager.php
@@ -13,6 +13,8 @@ namespace Hyperf\Kafka;
 
 use Hyperf\Contract\ConfigInterface;
 use Hyperf\Contract\StdoutLoggerInterface;
+use Hyperf\Coordinator\Constants;
+use Hyperf\Coordinator\CoordinatorManager;
 use Hyperf\Di\Annotation\AnnotationCollector;
 use Hyperf\Kafka\Annotation\Consumer as ConsumerAnnotation;
 use Hyperf\Kafka\Event\AfterConsume;
@@ -25,10 +27,10 @@ use longlang\phpkafka\Client\SwooleClient;
 use longlang\phpkafka\Consumer\ConsumeMessage;
 use longlang\phpkafka\Consumer\Consumer as LongLangConsumer;
 use longlang\phpkafka\Consumer\ConsumerConfig;
-use longlang\phpkafka\Exception\KafkaErrorException;
 use longlang\phpkafka\Socket\SwooleSocket;
 use Psr\Container\ContainerInterface;
 use Psr\EventDispatcher\EventDispatcherInterface;
+use Throwable;
 
 class ConsumerManager
 {
@@ -48,7 +50,7 @@ class ConsumerManager
          */
         foreach ($classes as $class => $annotation) {
             $instance = make($class);
-            if (! $instance instanceof AbstractConsumer || ! $annotation->enable) {
+            if (! $instance instanceof AbstractConsumer || ! $instance->isEnable($annotation->enable)) {
                 continue;
             }
 
@@ -97,46 +99,55 @@ class ConsumerManager
                 $longLangConsumer = new LongLangConsumer(
                     $consumerConfig,
                     function (ConsumeMessage $message) use ($consumer, $consumerConfig) {
-                        $this->dispatcher && $this->dispatcher->dispatch(new BeforeConsume($consumer, $message));
+                        $config = $this->getConfig();
+                        wait(function () use ($consumer, $consumerConfig, $message) {
+                            $this->dispatcher?->dispatch(new BeforeConsume($consumer, $message));
 
-                        $result = $consumer->consume($message);
+                            $result = $consumer->consume($message);
 
-                        if (! $consumerConfig->getAutoCommit()) {
-                            if (! is_string($result)) {
-                                throw new InvalidConsumeResultException('The result is invalid.');
+                            if (! $consumerConfig->getAutoCommit()) {
+                                if (! is_string($result)) {
+                                    throw new InvalidConsumeResultException('The result is invalid.');
+                                }
+
+                                if ($result === Result::ACK) {
+                                    $message->getConsumer()->ack($message);
+                                }
+
+                                if ($result === Result::REQUEUE) {
+                                    $this->producer->send($message->getTopic(), $message->getValue(), $message->getKey(), $message->getHeaders());
+                                }
                             }
 
-                            if ($result === Result::ACK) {
-                                $message->getConsumer()->ack($message);
-                            }
-
-                            if ($result === Result::REQUEUE) {
-                                $this->producer->send($message->getTopic(), $message->getValue(), $message->getKey(), $message->getHeaders());
-                            }
-                        }
-
-                        $this->dispatcher && $this->dispatcher->dispatch(new AfterConsume($consumer, $message, $result));
+                            $this->dispatcher?->dispatch(new AfterConsume($consumer, $message, $result));
+                        }, $config['consume_timeout'] ?? -1);
                     }
                 );
 
-                retry(
-                    3,
-                    function () use ($longLangConsumer) {
-                        try {
-                            $longLangConsumer->start();
-                        } catch (KafkaErrorException $exception) {
-                            $this->stdoutLogger->error($exception->getMessage());
-
-                            $this->dispatcher && $this->dispatcher->dispatch(new FailToConsume($this->consumer, [], $exception));
+                while (true) {
+                    try {
+                        if (CoordinatorManager::until(Constants::WORKER_EXIT)->yield(10)) {
+                            break;
                         }
-                    },
-                    10
-                );
+
+                        $longLangConsumer->start();
+                    } catch (Throwable $exception) {
+                        $this->stdoutLogger->warning((string) $exception);
+                        $this->dispatcher?->dispatch(new FailToConsume($this->consumer, [], $exception));
+                    }
+                }
+
+                $longLangConsumer->close();
+            }
+
+            public function getConfig(): array
+            {
+                return $this->config->get('kafka.' . $this->consumer->getPool());
             }
 
             public function getConsumerConfig(): ConsumerConfig
             {
-                $config = $this->config->get('kafka.' . $this->consumer->getPool());
+                $config = $this->getConfig();
                 $consumerConfig = new ConsumerConfig();
                 $consumerConfig->setAutoCommit($this->consumer->isAutoCommit());
                 $consumerConfig->setRackId($config['rack_id']);