+++ src/rpc-client/src/AbstractServiceClient.php
@@ -66,7 +66,8 @@ abstract class AbstractServiceClient
         $transporter = $protocol->getTransporter()->setLoadBalancer($loadBalancer);
         $this->client = make(Client::class)
             ->setPacker($protocol->getPacker())
-            ->setTransporter($transporter);
+            ->setTransporter($transporter)
+            ->setNormalizer($protocol->getNormalizer());
         $this->idGenerator = $this->getIdGenerator();
         $this->pathGenerator = $protocol->getPathGenerator();
         $this->dataFormatter = $protocol->getDataFormatter();